Added helper to call scoped command
